requirements.txt
# BE
fastapi
uvicorn[standard]

mysql-connector-python
python-jose[cryptography]
datetime
passlib[bcrypt]
python-multipart
sqlalchemy
pymysql

openai
uuid
minio

celery
kombu
ffmpeg-python